+++ platform/jvm/gradle-test-app/src/main/java/io/bitdrift/gradletestapp/diagnostics/fatalissues/CrashSdkInitializer.kt
@@ -40,6 +40,19 @@ object CrashSdkInitializer {
                             }
                         }
                     Bugsnag.start(application, bugsnagConfiguration)
+
+                    val bugsnagLastRunInfo = Bugsnag.getLastRunInfo()
+                    bugsnagLastRunInfo?.let { lastRunInfo ->
+                        Capture.Logger.logInfo(
+                            mapOf("crashed" to lastRunInfo.crashed.toString(),
+                                "crashedDuringLaunch" to lastRunInfo.crashedDuringLaunch.toString(),
+                                "consecutiveLaunchCrashes" to lastRunInfo.consecutiveLaunchCrashes.toString()
+                            )
+                        ) {
+                            "Bugsnag LastRunInfo"
+                        }
+                    }
+
                 }
             Timber.i("Bugsnag.start() took ${bugSnagStartTime.inWholeMilliseconds} ms")
             reportNonFatalIssue()